+++ src/components/SearchBar/searchBarChangelogData.ts
@@ -13,8 +13,9 @@ export const changelogData: ChangelogData[] = [
     date: "Prerelease",
     version: "Prerelease",
     type: "Update",
-    affects: ["Functionality"],
+    affects: ["Accessibility", "Documentation", "Functionality"],
     notes: [
+      "Removes role=search on the component wrapper.",
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
     ],
   },